backend/requirements.txt

fastapi>=0.109.0
uvicorn>=0.27.0

pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1

python-multipart>=0.0.7
httpx>=0.26.0

faiss-cpu>=1.7.4
google-genai>=1.0.0